requirements.txt
alabaster==0.7.8
alembic==1.7.7
amqp==5.1.0
aniso8601==7.0.0
anyio==3.5.0
appdirs==1.4.3
asgiref==3.5.0
asttokens==2.0.5
async-timeout==4.0.2
asyncio==3.4.3
asyncpg==0.25.0
attrs==21.4.0
Automat==0.8.0
Babel==2.6.0
backcall==0.2.0
backports.cached-property==1.0.1
billiard==3.6.4.0
blinker==1.4
cached-property==1.5.1
cassandra-driver==3.25.0
celery==5.2.3
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloud-init==20.1
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
decorator==5.1.1
Deprecated==1.2.13
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==2.2.1
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.16
email-validator==1.1.3
entrypoints==0.3
executing==0.8.3
fastapi==0.75.0
feedparser==5.2.1
filelock==3.0.12
Flask==1.1.1
geomet==0.2.1.post1
gitsome==0.8.0
graphql-core==3.2.0
graphql-relay==3.1.0
greenlet==1.1.2
h11==0.13.0
httplib2==0.14.0
httptools==0.2.0
humanize==1.0.0
hyperlink==19.0.0
idna==3.3
imagesize==1.2.0
importlib-metadata==1.5.0
incremental==16.10.1
iniconfig==1.1.1
install==1.3.5
invoke==1.3.0
ipython==8.1.1
itsdangerous==2.1.1
jedi==0.18.1
Jinja2==3.0.3
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
kombu==5.2.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
Mako==1.2.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
more-itertools==4.2.0
multidict==6.0.2
netifaces==0.10.4
numpydoc==0.7.0
oauthlib==3.1.0
olefile==0.46
orjson==3.6.7
packaging==21.3
parso==0.8.3
pbr==5.8.1
pexpect==4.8.0
pgspecial==1.9.0
pickleshare==0.7.5
Pillow==7.0.0
pluggy==1.0.0
ply==3.11
promise==2.3
prompt-toolkit==3.0.28
psycopg2==2.8.4
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.21.0
pydantic==1.9.0
Pygments==2.11.2
PyGObject==3.36.0
PyHamcrest==1.9.0
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.7
pyrsistent==0.15.5
pyserial==3.4
pytest==7.1.0
pytest-asyncio==0.18.2
python-apt==2.0.0
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==0.19.2
python-multipart==0.0.5
pytz==2022.1
PyYAML==5.4.1
redis==4.2.0
requests==2.27.1
requests-unixsocket==0.2.0
roman==2.0.0
Rx==1.6.1
SecretStorage==2.3.1
service-identity==18.1.0
setproctitle==1.1.10
simplejson==3.16.0
singledispatch==3.7.0
six==1.16.0
sniffio==1.2.0
Sphinx==1.8.5
SQLAlchemy==1.4.32
SQLAlchemy-Utils==0.38.2
sqlalchemy2-stubs==0.0.2a20
sqlmodel==0.0.6
sqlparse==0.2.4
ssh-import-id==5.10
stack-data==0.2.0
starlette==0.17.1
strawberry-graphql==0.102.2
systemd-python==234
tabulate==0.8.6
terminaltables==3.1.0
testresources==2.0.1
texttable==1.6.2
tomli==2.0.1
traitlets==5.1.1
Twisted==18.9.0
typing_extensions==4.1.1
ubuntu-advantage-tools==20.3
ufw==0.36
ujson==4.3.0
unattended-upgrades==0.1
urllib3==1.26.8
uvicorn==0.15.0
uvloop==0.16.0
vcrpy==4.1.1
vine==5.0.0
wadllib==1.3.3
watchgod==0.8
wcwidth==0.2.5
websocket-client==0.53.0
websockets==10.2
Werkzeug==0.16.1
wrapt==1.14.0
xonsh==0.9.13
yarl==1.7.2
zipp==1.0.0
zope.interface==4.7.1
